Revert "fix: import checkpoint instrumentation to avoid 64KB command limit (#…"
This reverts commit 4638c5d43bd7e74e518879035035c4f8786f4092.
