--- kubeflow/trainer/rhai/instrumentation/__init__.py
@@ -1,1 +1,0 @@
-"""RHAI instrumentation modules for training pods."""
