+++ kubeflow/trainer/rhai/transformers_test.py
@@ -263,31 +263,29 @@ def test_func_callable_validation(test_case):
     "test_case",
     [
         TestCase(
-            name="basic generation - returns string with module import",
+            name="basic generation - returns string",
             expected_status="success",
             config={"port": 28080},
             expected_output=[
                 ("isinstance", str),
-                ("import kubeflow as runtime_kubeflow", True),
-                (
-                    "from kubeflow.trainer.rhai.instrumentation.progression import create_progression_instrumentation",
-                    True,
-                ),
-                ("apply_progression_tracking", True),
+                ("class KubeflowProgressCallback", True),
+                ("class ProgressionMetricsHandler", True),
+                ("def apply_progression_tracking", True),
                 ("{{user_func_import_and_call}}", True),
                 ("metrics_port=28080", True),
             ],
         ),
         TestCase(
-            name="module import - SDK version checking",
+            name="self-contained - no SDK imports",
             expected_status="success",
             config={"port": 28080},
             expected_output=[
-                ("import kubeflow", True),
-                ("CLIENT_SDK_VERSION", True),
-                ("RUNTIME_SDK_VERSION", True),
-                ("MIN_SDK_VERSION", True),
-                ("version.parse", True),
+                ("from kubeflow", False),
+                ("import kubeflow", False),
+                ("class ProgressionMetricsHandler", True),
+                ("class KubeflowProgressCallback", True),
+                ("def apply_progression_tracking", True),
+                ("from transformers import", True),
             ],
         ),
         TestCase(
@@ -302,25 +300,66 @@ def test_func_callable_validation(test_case):
             ],
         ),
         TestCase(
-            name="error handling - version compatibility",
+            name="completeness - all callback methods",
             expected_status="success",
             config={"port": 28080},
             expected_output=[
-                ("if not client_version_valid or not runtime_version_valid:", True),
-                ("packages_to_install", True),
-                ("clusterTrainingRuntimes", True),
-                ("RuntimeError", True),
+                ("def on_train_begin", True),
+                ("def on_step_end", True),
+                ("def on_log", True),
+                ("def on_train_end", True),
+                ("_original_init", True),
+                ("def _instrumented_trainer_init", True),
             ],
         ),
         TestCase(
-            name="initialization messages",
+            name="implementation details - progress tracking logic",
             expected_status="success",
             config={"port": 28080},
             expected_output=[
-                ("[Kubeflow] Initializing progression tracking", True),
-                ("[Kubeflow] Progression tracking enabled", True),
-                ("Client SDK", True),
-                ("Runtime SDK", True),
+                ("from kubeflow.trainer.rhai", False),
+                ("state.global_step", True),
+                ("progress_pct", True),
+                ("elapsed_sec", True),
+            ],
+        ),
+        TestCase(
+            name="dataclass - ProgressionMetricsState",
+            expected_status="success",
+            config={"port": 28080},
+            expected_output=[
+                ("@dataclass", True),
+                ("class ProgressionMetricsState", True),
+                ("progressPercentage", True),
+                ("estimatedRemainingSeconds", True),
+                ("currentStep", True),
+                ("totalSteps", True),
+                ("trainMetrics", True),
+                ("evalMetrics", True),
+                ("asdict", True),
+            ],
+        ),
+        TestCase(
+            name="metrics state initialization",
+            expected_status="success",
+            config={"port": 28080},
+            expected_output=[
+                ("progressPercentage: int | None = None", True),
+                ("currentStep: int = 0", True),
+                ("currentEpoch: float = 0.0", True),
+                ("trainMetrics: dict[str, Any] = field(default_factory=dict)", True),
+                ("evalMetrics: dict[str, Any] = field(default_factory=dict)", True),
+            ],
+        ),
+        TestCase(
+            name="thread safety mechanisms",
+            expected_status="success",
+            config={"port": 28080},
+            expected_output=[
+                ("_progression_metrics_lock", True),
+                ("threading.Lock()", True),
+                ("with _progression_metrics_lock:", True),
+                ("_update_progression_metrics", True),
             ],
         ),
     ],
@@ -752,10 +791,21 @@ def test_callback_completion_state_protection(test_case):
     sys.modules["transformers"] = mock_transformers
 
     try:
-        # Import progression instrumentation directly from module
-        from kubeflow.trainer.rhai.instrumentation.progression import (
-            create_progression_instrumentation,
-        )
+        wrapper = get_transformers_instrumentation_wrapper(metrics_port=28080)
+        namespace = {}
+        # Replace user code placeholder
+        wrapper_code = wrapper.replace("{{user_func_import_and_call}}", "pass")
+        # Skip the actual patching call (line after _create_progression_instrumentation)
+        lines = wrapper_code.split("\n")
+        modified_lines = []
+        for line in lines:
+            # Comment out the standalone apply_progression_tracking() call
+            if line.strip() == "apply_progression_tracking()":
+                modified_lines.append("# apply_progression_tracking()  # Skipped in tests")
+            else:
+                modified_lines.append(line)
+        wrapper_code = "\n".join(modified_lines)
+        exec(wrapper_code, namespace)
 
         (
             _,
@@ -763,7 +813,7 @@ def test_callback_completion_state_protection(test_case):
             _,
             get_metrics_json,
             _,
-        ) = create_progression_instrumentation(28080)
+        ) = namespace["_create_progression_instrumentation"](28080)
         callback = callback_class(metrics_port=28080)
 
         import json
@@ -899,10 +949,20 @@ def test_completion_detection_via_on_step_end(test_case):
     sys.modules["transformers"] = mock_transformers
 
     try:
-        # Import progression instrumentation directly from module
-        from kubeflow.trainer.rhai.instrumentation.progression import (
-            create_progression_instrumentation,
-        )
+        wrapper = get_transformers_instrumentation_wrapper(metrics_port=28080)
+        namespace = {}
+        # Replace user code placeholder
+        wrapper_code = wrapper.replace("{{user_func_import_and_call}}", "pass")
+        # Skip the actual patching call
+        lines = wrapper_code.split("\n")
+        modified_lines = []
+        for line in lines:
+            if line.strip() == "apply_progression_tracking()":
+                modified_lines.append("# apply_progression_tracking()  # Skipped in tests")
+            else:
+                modified_lines.append(line)
+        wrapper_code = "\n".join(modified_lines)
+        exec(wrapper_code, namespace)
 
         (
             _,
@@ -910,7 +970,7 @@ def test_completion_detection_via_on_step_end(test_case):
             _,
             get_metrics_json,
             _,
-        ) = create_progression_instrumentation(28080)
+        ) = namespace["_create_progression_instrumentation"](28080)
         callback = callback_class(metrics_port=28080)
 
         import json
@@ -1039,10 +1099,21 @@ def test_callback_metrics_categorization(test_case):
     sys.modules["transformers"] = mock_transformers
 
     try:
-        # Import progression instrumentation directly from module
-        from kubeflow.trainer.rhai.instrumentation.progression import (
-            create_progression_instrumentation,
-        )
+        wrapper = get_transformers_instrumentation_wrapper(metrics_port=28080)
+        namespace = {}
+        # Replace user code placeholder
+        wrapper_code = wrapper.replace("{{user_func_import_and_call}}", "pass")
+        # Skip the actual patching call (line after _create_progression_instrumentation)
+        lines = wrapper_code.split("\n")
+        modified_lines = []
+        for line in lines:
+            # Comment out the standalone apply_progression_tracking() call
+            if line.strip() == "apply_progression_tracking()":
+                modified_lines.append("# apply_progression_tracking()  # Skipped in tests")
+            else:
+                modified_lines.append(line)
+        wrapper_code = "\n".join(modified_lines)
+        exec(wrapper_code, namespace)
 
         (
             _,
@@ -1050,7 +1121,7 @@ def test_callback_metrics_categorization(test_case):
             _,
             get_metrics_json,
             _,
-        ) = create_progression_instrumentation(28080)
+        ) = namespace["_create_progression_instrumentation"](28080)
         callback = callback_class(metrics_port=28080)
 
         import json
@@ -1176,10 +1247,18 @@ def test_honest_progress_reporting(test_case):
     sys.modules["transformers"] = mock_transformers
 
     try:
-        # Import progression instrumentation directly from module
-        from kubeflow.trainer.rhai.instrumentation.progression import (
-            create_progression_instrumentation,
-        )
+        wrapper = get_transformers_instrumentation_wrapper(metrics_port=28080)
+        namespace = {}
+        wrapper_code = wrapper.replace("{{user_func_import_and_call}}", "pass")
+        lines = wrapper_code.split("\n")
+        modified_lines = []
+        for line in lines:
+            if line.strip() == "apply_progression_tracking()":
+                modified_lines.append("# apply_progression_tracking()  # Skipped in tests")
+            else:
+                modified_lines.append(line)
+        wrapper_code = "\n".join(modified_lines)
+        exec(wrapper_code, namespace)
 
         (
             _,
@@ -1187,7 +1266,7 @@ def test_honest_progress_reporting(test_case):
             _,
             get_metrics_json,
             _,
-        ) = create_progression_instrumentation(28080)
+        ) = namespace["_create_progression_instrumentation"](28080)
         callback = callback_class(metrics_port=28080)
 
         import json
@@ -1999,8 +2078,8 @@ def test_find_latest_checkpoint_none_output_dir():
 
 
 def test_auto_resume_code_generation():
-    """Test that checkpoint instrumentation is imported (not inlined)."""
-    print("Executing test: checkpoint instrumentation import")
+    """Test that auto-resume code is generated in checkpoint injection."""
+    print("Executing test: auto-resume code generation")
 
     from kubeflow.trainer.rhai.transformers import get_jit_checkpoint_injection_code
 
@@ -2010,24 +2089,71 @@ def test_auto_resume_code_generation():
         enable_jit_checkpoint=True,
     )
 
-    # Verify checkpoint instrumentation is imported from package (not inlined)
-    assert (
-        "from kubeflow.trainer.rhai.instrumentation.checkpoint import create_checkpoint_instrumentation"
-        in checkpoint_header
-    ), "Should import checkpoint instrumentation from package"
-    assert "create_checkpoint_instrumentation(checkpoint_config)" in checkpoint_header, (
-        "Should call create_checkpoint_instrumentation"
+    # Verify auto-resume logic is present in generated code
+    assert "_find_latest_checkpoint" in checkpoint_header, "Should include _find_latest_checkpoint"
+    assert "def _patched_train" in checkpoint_header, "Should include _patched_train"
+    assert "resume_from_checkpoint" in checkpoint_header, (
+        "Should include resume_from_checkpoint logic"
     )
-    assert "apply_checkpointing()" in checkpoint_header, "Should apply checkpointing"
+    assert "if resume_from_checkpoint is None" in checkpoint_header, "Should check if user set it"
+    assert "Auto-resuming from:" in checkpoint_header, "Should log auto-resume action"
+    assert "self.train = _patched_train" in checkpoint_header, "Should patch train method"
 
-    # Verify we're NOT inlining the code (old approach)
-    assert "_find_latest_checkpoint" not in checkpoint_header, (
-        "Should NOT inline checkpoint code (import instead)"
-    )
-    assert "import re" not in checkpoint_header, "Should NOT inline imports (import instead)"
-    assert "import shutil" not in checkpoint_header, "Should NOT inline imports (import instead)"
+    # Verify imports needed for auto-resume
+    assert "import re" in checkpoint_header, "Should import re for regex"
+    assert "import shutil" in checkpoint_header, "Should import shutil for rmtree"
+    assert "import os" in checkpoint_header, "Should import os"
 
     print("test execution complete")
+
+
+def test_auto_resume_user_override():
+    """Test that user's explicit resume_from_checkpoint is not overridden."""
+    print("Executing test: auto-resume respects user override")
+
+    import sys
+    from unittest.mock import Mock
+
+    # Mock transformers module
+    mock_transformers = Mock()
+    mock_transformers.TrainerCallback = type("TrainerCallback", (), {})
+    mock_transformers.trainer = Mock()
+    mock_transformers.trainer_utils = Mock()
+    mock_transformers.trainer_utils.PREFIX_CHECKPOINT_DIR = "checkpoint"
+    sys.modules["transformers"] = mock_transformers
+    sys.modules["transformers.trainer_utils"] = mock_transformers.trainer_utils
+
+    # Mock torch module
+    mock_torch = Mock()
+    mock_torch.cuda.is_available.return_value = False
+    sys.modules["torch"] = mock_torch
+
+    try:
+        from kubeflow.trainer.rhai.transformers import get_jit_checkpoint_injection_code
+
+        checkpoint_header, _ = get_jit_checkpoint_injection_code(
+            output_dir="/tmp/checkpoints",
+            periodic_checkpoint_config=None,
+            enable_jit_checkpoint=True,
+        )
+
+        # Verify the logic: only auto-resume if resume_from_checkpoint is None
+        assert "if resume_from_checkpoint is None" in checkpoint_header
+
+        assert (
+            "resume_from_checkpoint is None and training_args" in checkpoint_header
+            or "if resume_from_checkpoint is None" in checkpoint_header
+        )
+
+        print("test execution complete")
+
+    finally:
+        if "transformers" in sys.modules:
+            del sys.modules["transformers"]
+        if "transformers.trainer_utils" in sys.modules:
+            del sys.modules["transformers.trainer_utils"]
+        if "torch" in sys.modules:
+            del sys.modules["torch"]
 
 
 def test_transformers_trainer_s3_requires_data_connection_name():
@@ -2255,6 +2381,228 @@ def test_save_on_each_node_validation_error(tmp_path):
 
     assert returncode == 0
     assert "save_on_each_node=True is not supported when output_dir is an S3 URI" in stdout
+
+    print("test execution complete")
+
+
+def test_async_upload_worker_scaffolding():
+    """Test async upload worker scaffolding is present in generated code."""
+    print("Executing test: async upload worker scaffolding")
+
+    checkpoint_header, _ = get_jit_checkpoint_injection_code(
+        output_dir="/mnt/kubeflow-checkpoints",
+        cloud_remote_storage_uri="s3://my-bucket/model-checkpoints",
+        enable_jit_checkpoint=True,
+    )
+
+    assert "LifoQueue" in checkpoint_header
+    assert "daemon=False" in checkpoint_header
+    assert "KubeflowCheckpointUploader" in checkpoint_header
+    assert "1 hour" in checkpoint_header
+    assert "self._upload_thread.join(timeout=" in checkpoint_header
+
+    print("test execution complete")
+
+
+def test_async_parallel_upload_scaffolding():
+    """Test parallel upload scaffolding is present in generated code."""
+    print("Executing test: async parallel upload scaffolding")
+
+    checkpoint_header, _ = get_jit_checkpoint_injection_code(
+        output_dir="/mnt/kubeflow-checkpoints",
+        cloud_remote_storage_uri="s3://my-bucket/model-checkpoints",
+        enable_jit_checkpoint=True,
+    )
+
+    assert "ThreadPoolExecutor" in checkpoint_header
+    assert "as_completed" in checkpoint_header
+    assert "_parallel_upload_files" in checkpoint_header
+
+    print("test execution complete")
+
+
+def test_async_upload_execution(tmp_path):
+    """Integration test: async upload runs and cleans staging."""
+    print("Executing test: async upload execution")
+
+    import subprocess
+    import sys
+
+    from kubeflow.trainer.rhai.constants import CHECKPOINT_INCOMPLETE_MARKER
+
+    output_dir = tmp_path / "checkpoints"
+
+    checkpoint_header, _ = get_jit_checkpoint_injection_code(
+        output_dir=str(output_dir),
+        cloud_remote_storage_uri="s3://test-bucket/model-checkpoints",
+        enable_jit_checkpoint=True,
+    )
+    checkpoint_code = checkpoint_header
+
+    fsspec_stub = """
+class MockS3FileSystem:
+    def __init__(self, *args, **kwargs):
+        pass
+
+    def ls(self, path, detail=False):
+        return []
+
+    def exists(self, path):
+        return False
+
+    def pipe(self, path, data):
+        print(f"PIPE={path}")
+
+    def cat(self, path):
+        return b"test"
+
+    def put_file(self, local, remote):
+        print(f"PUT_FILE={remote}")
+
+    def rm_file(self, path):
+        print(f"RM_FILE={path}")
+
+    def get(self, src, dst, recursive=False, callback=None):
+        pass
+
+    def du(self, path, total=True, maxdepth=None):
+        return 0
+
+def filesystem(protocol, **kwargs):
+    return MockS3FileSystem()
+"""
+
+    torch_stub = """
+class distributed:
+    @staticmethod
+    def is_available():
+        return False
+
+    @staticmethod
+    def is_initialized():
+        return False
+
+    @staticmethod
+    def barrier():
+        pass
+
+class cuda:
+    @staticmethod
+    def is_available():
+        return False
+"""
+
+    transformers_stub = """
+class TrainerCallback:
+    pass
+
+class trainer_utils:
+    PREFIX_CHECKPOINT_DIR = "checkpoint"
+
+PREFIX_CHECKPOINT_DIR = "checkpoint"
+
+class TrainerState:
+    def __init__(self):
+        self.global_step = 0
+        self.is_world_process_zero = True
+
+class Trainer:
+    def __init__(self, *args, **kwargs):
+        self.state = TrainerState()
+        self.model = None
+        self._init_args = args
+        self._init_kwargs = kwargs
+
+    def train(self, resume_from_checkpoint=None):
+        return {"train_called": True, "resume_from": resume_from_checkpoint}
+"""
+
+    test_file = tmp_path / "test_async_upload_execution.py"
+
+    test_code = f"""
+import os
+import sys
+import types
+
+fsspec_module = types.ModuleType('fsspec')
+exec('''{fsspec_stub}''', fsspec_module.__dict__)
+sys.modules['fsspec'] = fsspec_module
+
+torch_module = types.ModuleType('torch')
+exec('''{torch_stub}''', torch_module.__dict__)
+sys.modules['torch'] = torch_module
+sys.modules['torch.distributed'] = torch_module.distributed
+
+transformers_module = types.ModuleType('transformers')
+exec('''{transformers_stub}''', transformers_module.__dict__)
+sys.modules['transformers'] = transformers_module
+
+trainer_utils_module = types.ModuleType('transformers.trainer_utils')
+trainer_utils_module.PREFIX_CHECKPOINT_DIR = "checkpoint"
+sys.modules['transformers.trainer_utils'] = trainer_utils_module
+
+{checkpoint_code}
+
+from transformers import TrainerCallback
+
+callback_class = None
+for name, obj in list(globals().items()):
+    if isinstance(obj, type) and issubclass(obj, TrainerCallback) and name != 'TrainerCallback':
+        callback_class = obj
+        break
+
+if not callback_class:
+    print("ERROR: JITCheckpointCallback not found")
+    sys.exit(1)
+
+callback = callback_class(cloud_remote_storage_uri="s3://test-bucket/model-checkpoints")
+
+class MockArgs:
+    output_dir = r"{output_dir}"
+    local_process_index = 0
+
+class MockState:
+    global_step = 3
+    is_world_process_zero = True
+
+checkpoint_path = os.path.join(MockArgs.output_dir, "checkpoint-3")
+os.makedirs(checkpoint_path, exist_ok=True)
+with open(os.path.join(checkpoint_path, "model.bin"), "w", encoding="utf-8") as f:
+    f.write("data")
+
+callback.on_save(MockArgs(), MockState(), None)
+callback.shutdown_upload_worker()
+
+if True:
+    staging_checkpoint = os.path.join(
+        MockArgs.output_dir, CHECKPOINT_STAGING_DIR, "checkpoint-3"
+    )
+    print(f"STAGING_CHECKPOINT_EXISTS={{os.path.exists(staging_checkpoint)}}")
+    print(f"CHECKPOINT_EXISTS={{os.path.exists(checkpoint_path)}}")
+
+print("TEST_COMPLETE=True")
+"""
+
+    test_file.write_text(test_code)
+
+    result = subprocess.run(
+        [sys.executable, str(test_file)], capture_output=True, text=True, timeout=10
+    )
+
+    output = result.stdout
+    print(f"Test output:\\n{output}")
+
+    if result.returncode != 0:
+        print(f"Test stderr:\\n{result.stderr}")
+
+    assert result.returncode == 0, f"Execution failed with return code {result.returncode}"
+    assert "TEST_COMPLETE=True" in output
+    expected_marker = f"{CHECKPOINT_INCOMPLETE_MARKER}.node-0-rank-0"
+    assert f"PIPE=checkpoint-3/{expected_marker}" in output
+    assert "PUT_FILE=checkpoint-3/model.bin" in output
+    assert f"RM_FILE=checkpoint-3/{expected_marker}" in output
+    assert "STAGING_CHECKPOINT_EXISTS=False" in output
+    assert "CHECKPOINT_EXISTS=False" in output
 
     print("test execution complete")
 
@@ -2651,6 +2999,35 @@ def test_output_dir_normalization(test_case):
         assert type(e) is test_case.expected_error
 
     print("test execution complete")
+
+
+def test_s3_access_retry_code_generation():
+    """Test that S3 access verification includes 3-retry logic in generated code."""
+    print("Executing test: S3 retry logic in generated code")
+
+    # Generate checkpoint injection code with S3 config
+    checkpoint_header, _ = get_jit_checkpoint_injection_code(
+        output_dir="/mnt/checkpoints",
+        cloud_remote_storage_uri="s3://test-bucket/path",
+        enable_jit_checkpoint=True,
+        verify_cloud_storage_access=True,
+    )
+    code = checkpoint_header
+
+    # Verify retry logic is present in generated code
+    assert "for attempt in range(1, 4):" in code, "3-attempt retry loop not found"
+    assert "last_error = None" in code, "Error tracking not found"
+    assert "last_error = e" in code, "Error capture not found"
+    assert "if attempt < 3:" in code, "Retry condition not found"
+    assert "time.sleep(1)" in code, "Backoff sleep not found"
+    assert "if last_error:" in code, "Error re-raise check not found"
+    assert "raise last_error" in code, "Error propagation not found"
+
+    # Verify code structure contains the access verification section
+    assert "verify_cloud_storage_access" in code
+    assert 'self.remote_fs.pipe(test_file, b"test")' in code
+    assert "self.remote_fs.cat(test_file)" in code
+    assert "self.remote_fs.rm_file(test_file)" in code
 
 
 # ============================================================================
@@ -3912,236 +4289,6 @@ print("TEST_COMPLETE=True")
     # Sentinel removed AFTER save (checkpoint is resume-ready)
     assert "SENTINEL_REMOVED_AFTER_SAVE=True" in output
     assert "TEST_COMPLETE=True" in output
-
-    print("test execution complete")
-
-
-def test_checkpoint_error_handling_module_not_found(tmp_path):
-    """Test error handling when runtime SDK version is below minimum requirement."""
-    print("Executing test: checkpoint error handling - runtime SDK version too old")
-
-    import subprocess
-    import sys
-
-    # Generate checkpoint code
-    checkpoint_header, _ = get_jit_checkpoint_injection_code(
-        output_dir="/tmp/checkpoints",
-        enable_jit_checkpoint=True,
-    )
-
-    # Write checkpoint header to a separate file
-    checkpoint_file = tmp_path / "checkpoint_code.py"
-    checkpoint_file.write_text(checkpoint_header)
-
-    # Create test file that simulates old runtime SDK version
-    test_file = tmp_path / "test_old_runtime_version.py"
-    test_code = f"""
-import sys
-
-# Mock old runtime kubeflow version before executing checkpoint code
-from unittest.mock import Mock
-mock_kubeflow = Mock()
-mock_kubeflow.__version__ = "0.3.0"  # Below minimum 0.4.0
-sys.modules['kubeflow'] = mock_kubeflow
-
-try:
-    # Execute checkpoint code
-    with open('{checkpoint_file}', 'r') as f:
-        exec(f.read())
-    print("ERROR: Should have raised RuntimeError")
-    sys.exit(1)
-except RuntimeError as e:
-    error_msg = str(e)
-    # Verify error message content
-    assert "Checkpoint instrumentation requires SDK version" in error_msg, f"Wrong error: {{error_msg}}"
-    assert "Runtime SDK version 0.3.0 is below minimum" in error_msg
-    assert "packages_to_install" in error_msg
-    assert "clusterTrainingRuntimes" in error_msg
-    print("SUCCESS: Correct error raised for old runtime SDK")
-    sys.exit(0)
-except Exception as e:
-    print(f"ERROR: Unexpected exception: {{type(e).__name__}}: {{e}}")
-    import traceback
-    traceback.print_exc()
-    sys.exit(1)
-"""
-
-    test_file.write_text(test_code)
-
-    result = subprocess.run(
-        [sys.executable, str(test_file)], capture_output=True, text=True, timeout=10
-    )
-
-    if result.returncode != 0:
-        print(f"Test output: {result.stdout}")
-        print(f"Test stderr: {result.stderr}")
-
-    assert result.returncode == 0, "Error handling test failed"
-    assert "SUCCESS: Correct error raised for old runtime SDK" in result.stdout
-
-    print("test execution complete")
-
-
-def test_checkpoint_error_handling_version_mismatch_warning(tmp_path):
-    """Test that different SDK versions are allowed as long as both are >= 0.4.0."""
-    print("Executing test: checkpoint with different but compatible SDK versions")
-
-    import subprocess
-    import sys
-
-    # Generate checkpoint code
-    checkpoint_header, _ = get_jit_checkpoint_injection_code(
-        output_dir="/tmp/checkpoints",
-        enable_jit_checkpoint=True,
-    )
-
-    # Write checkpoint header to a separate file
-    checkpoint_file = tmp_path / "checkpoint_code.py"
-    checkpoint_file.write_text(checkpoint_header)
-
-    # Mock modules with different version (but both >= 0.4.0)
-    test_file = tmp_path / "test_version_compatible.py"
-    test_code = f"""
-import sys
-from unittest.mock import Mock, patch
-
-# Mock transformers
-mock_transformers = Mock()
-mock_transformers.TrainerCallback = type("TrainerCallback", (), {{}})
-mock_transformers.trainer_utils = Mock()
-mock_transformers.trainer_utils.PREFIX_CHECKPOINT_DIR = "checkpoint"
-sys.modules['transformers'] = mock_transformers
-sys.modules['transformers.trainer_utils'] = mock_transformers.trainer_utils
-
-# Mock torch
-mock_torch = Mock()
-mock_torch.cuda.is_available.return_value = False
-mock_torch.distributed.is_available.return_value = False
-sys.modules['torch'] = mock_torch
-sys.modules['torch.distributed'] = mock_torch.distributed
-
-try:
-    # Import checkpoint instrumentation first
-    from kubeflow.trainer.rhai.instrumentation.checkpoint import create_checkpoint_instrumentation
-
-    # Mock create_checkpoint_instrumentation to avoid actual execution
-    def mock_create(*args, **kwargs):
-        return None, None, lambda: None, lambda: None
-
-    # Patch create_checkpoint_instrumentation and execute checkpoint code
-    with patch('kubeflow.trainer.rhai.instrumentation.checkpoint.create_checkpoint_instrumentation', side_effect=mock_create):
-        # Modify runtime version before executing checkpoint code
-        import kubeflow
-        kubeflow.__version__ = "0.5.0"  # Different from client version (but >= 0.4.0)
-
-        with open('{checkpoint_file}', 'r') as f:
-            exec(f.read())
-
-    # If we get here, no exception was raised - success!
-    print("SUCCESS: Different but compatible SDK versions work without error")
-    sys.exit(0)
-
-except Exception as e:
-    print(f"ERROR: {{e}}")
-    import traceback
-    traceback.print_exc()
-    sys.exit(1)
-"""
-
-    test_file.write_text(test_code)
-
-    result = subprocess.run(
-        [sys.executable, str(test_file)], capture_output=True, text=True, timeout=10
-    )
-
-    if result.returncode != 0:
-        print(f"Test output: {result.stdout}")
-        print(f"Test stderr: {result.stderr}")
-
-    assert result.returncode == 0, "Compatible version test failed"
-    assert "SUCCESS: Different but compatible SDK versions work without error" in result.stdout
-
-    # Verify warning message is shown for version mismatch
-    assert "[Kubeflow] Warning: This job was created with SDK" in result.stdout
-    assert "but the runtime image has SDK" in result.stdout
-    assert "If you encounter unexpected errors" in result.stdout
-
-    print("test execution complete")
-
-
-def test_checkpoint_error_handling_generic_exception(tmp_path):
-    """Test error handling for generic exceptions (config/credential errors)."""
-    print("Executing test: checkpoint error handling - generic exception")
-
-    import subprocess
-    import sys
-
-    # Generate checkpoint code
-    checkpoint_header, _ = get_jit_checkpoint_injection_code(
-        output_dir="/tmp/checkpoints",
-        enable_jit_checkpoint=True,
-    )
-
-    # Write checkpoint header to a separate file
-    checkpoint_file = tmp_path / "checkpoint_code.py"
-    checkpoint_file.write_text(checkpoint_header)
-
-    test_file = tmp_path / "test_generic_exception.py"
-    test_code = f"""
-import sys
-from unittest.mock import Mock, patch
-
-# Mock modules
-mock_transformers = Mock()
-mock_transformers.TrainerCallback = type("TrainerCallback", (), {{}})
-mock_transformers.trainer_utils = Mock()
-mock_transformers.trainer_utils.PREFIX_CHECKPOINT_DIR = "checkpoint"
-sys.modules['transformers'] = mock_transformers
-sys.modules['transformers.trainer_utils'] = mock_transformers.trainer_utils
-
-mock_torch = Mock()
-mock_torch.cuda.is_available.return_value = False
-mock_torch.distributed.is_available.return_value = False
-sys.modules['torch'] = mock_torch
-sys.modules['torch.distributed'] = mock_torch.distributed
-
-# Create a mock that raises a generic exception (e.g., invalid config)
-def mock_create_checkpoint(*args, **kwargs):
-    raise ValueError("Invalid configuration: output_dir cannot be empty")
-
-# Patch the import
-with patch('kubeflow.trainer.rhai.instrumentation.checkpoint.create_checkpoint_instrumentation', side_effect=mock_create_checkpoint):
-    try:
-        # Execute checkpoint code
-        with open('{checkpoint_file}', 'r') as f:
-            exec(f.read())
-        print("ERROR: Should have raised RuntimeError")
-        sys.exit(1)
-    except RuntimeError as e:
-        error_msg = str(e)
-        # Verify error message shows the actual error (not SDK versions)
-        assert "Failed to initialize checkpoint instrumentation" in error_msg
-        assert "ValueError" in error_msg
-        assert "Invalid configuration" in error_msg
-        # Should NOT mention SDK versions (already verified as compatible)
-        assert "Client SDK:" not in error_msg
-        assert "Runtime SDK:" not in error_msg
-        print("SUCCESS: Generic exception handled correctly")
-        sys.exit(0)
-"""
-
-    test_file.write_text(test_code)
-
-    result = subprocess.run(
-        [sys.executable, str(test_file)], capture_output=True, text=True, timeout=10
-    )
-
-    if result.returncode != 0:
-        print(f"Test output: {result.stdout}")
-        print(f"Test stderr: {result.stderr}")
-
-    assert result.returncode == 0, "Generic exception test failed"
-    assert "SUCCESS: Generic exception handled correctly" in result.stdout
 
     print("test execution complete")
